+++ src/core/commands.py
@@ -20,6 +20,8 @@
 # along with pashmak.  If not, see <https://www.gnu.org/licenses/>.
 ##################################################
 
+''' Only some aliases for operations '''
+
 # operations imports
 from operations import set as op_set
 from operations import free as op_free
@@ -51,86 +53,116 @@ from operations import endnamespace as op_endnamespace
 from operations import use as op_use
 
 class Commands:
+    ''' Only some aliases for operations '''
+
     def run_set(self, op: dict):
+        ''' run set '''
         op_set.run(self, op)
 
     def run_free(self, op: dict):
+        ''' run free '''
         op_free.run(self, op)
 
     def run_copy(self, op: dict):
+        ''' run copy '''
         op_copy.run(self, op)
 
     def run_out(self, op: dict):
+        ''' run out '''
         op_out.run(self, op)
 
     def run_read(self, op: dict):
+        ''' run read '''
         op_read.run(self, op)
 
     def run_return(self, op: dict):
+        ''' run return '''
         op_return.run(self, op)
 
     def run_func(self, op: dict):
+        ''' run func '''
         op_func.run(self, op)
 
     def run_endfunc(self, op: dict):
+        ''' run endfunc '''
         op_endfunc.run(self, op)
 
     def run_required(self, op: dict):
+        ''' run required '''
         op_required.run(self, op)
 
     def run_typeof(self, op: dict):
+        ''' run typeof '''
         op_typeof.run(self, op)
 
     def run_system(self, op: dict):
+        ''' run system '''
         op_system.run(self, op)
 
     def run_include(self, op: dict):
+        ''' run include '''
         op_include.run(self, op)
 
     def run_goto(self, op: dict):
+        ''' run goto '''
         op_goto.run(self, op)
 
     def run_gotoif(self, op: dict):
+        ''' run gotoif '''
         op_gotoif.run(self, op)
 
     def run_fread(self, op: dict):
+        ''' run fread '''
         op_fread.run(self, op)
 
     def run_fwrite(self, op: dict):
+        ''' run fwrite '''
         op_fwrite.run(self, op)
 
     def run_chdir(self, op: dict):
+        ''' run chdir '''
         op_chdir.run(self, op)
 
     def run_cwd(self, op: dict):
+        ''' run cwd '''
         op_cwd.run(self, op)
 
     def run_isset(self, op: dict):
+        ''' run isset '''
         op_isset.run(self, op)
 
     def run_try(self, op: dict):
+        ''' run try '''
         op_try.run(self, op)
 
     def run_endtry(self, op: dict):
+        ''' run endtry '''
         op_endtry.run(self, op)
 
     def run_eval(self, op: dict):
+        ''' run eval '''
         op_eval.run(self, op)
 
     def run_arraypush(self, op: dict):
+        ''' run arraypush '''
         op_arraypush.run(self, op)
 
     def run_arraypop(self, op: dict):
+        ''' run arraypop '''
         op_arraypop.run(self, op)
 
     def run_python(self, op: dict):
+        ''' run python '''
         op_python.run(self, op)
 
     def run_namespace(self, op: dict):
+        ''' run namespace '''
         op_namespace.run(self, op)
 
     def run_endnamespace(self, op: dict):
+        ''' run endnamespace '''
         op_endnamespace.run(self, op)
 
     def run_use(self, op: dict):
+        ''' run use '''
         op_use.run(self, op)