+++ src/operations/copy.py
@@ -20,6 +20,8 @@
 # along with pashmak.  If not, see <https://www.gnu.org/licenses/>.
 ##################################################
 
+''' Copies a variable or mem into another variable '''
+
 def run(self, op: dict):
     ''' Copies a variable or mem into another variable '''
     args = op['args']
@@ -37,7 +39,7 @@ def run(self, op: dict):
         except:
             self.raise_variable_error(first_var, op)
 
-    if len(args[1]) == 0:
+    if not args[1]:
         self.raise_error('SyntaxError', 'one or more arguments are empty', op)
 
     second_var = args[1]
@@ -47,7 +49,7 @@ def run(self, op: dict):
             first_var_value = self.get_mem()
         else:
             first_var_value = self.get_var(first_var[1:])
-    except:
+    except KeyError:
         self.raise_variable_error(first_var, op)
 
     try: