+++ src/operations/include.py
@@ -20,6 +20,8 @@
 # along with pashmak.  If not, see <https://www.gnu.org/licenses/>.
 ##################################################
 
+''' Includes another script file to program '''
+
 import os
 from syntax import parser
 from core import modules
@@ -29,6 +31,7 @@ def run(self, op: dict):
 
     self.require_one_argument(op, 'include operation requires argument')
     arg = op['args'][0]
+    self.arg_should_be_variable_or_mem(arg, op)
 
     content = ''
 
@@ -38,7 +41,7 @@ def run(self, op: dict):
         self.variable_required(arg[1:], op)
         path = self.get_var(arg[1:])
 
-    if path == None:
+    if not path:
         return
 
     if path[0] == '@':
@@ -50,14 +53,16 @@ def run(self, op: dict):
                 self.included_modules.append(module_name)
             else:
                 return
-        except:
+        except KeyError:
             self.raise_error('ModuleError', 'undefined module "' + module_name + '"', op)
     else:
         if path[0] != '/':
-            path = os.path.dirname(self.main_filename) + '/' + path
+            path = os.path.dirname(os.path.abspath(self.main_filename)) + '/' + path
         try:
             content = open(path, 'r').read()
-        except Exception as ex:
+        except FileNotFoundError as ex:
+            self.raise_error('FileError', str(ex), op)
+        except PermissionError as ex:
             self.raise_error('FileError', str(ex), op)
 
     operations = parser.parse(content)