+++ tests/items/test_stdlib.py
@@ -69,31 +69,31 @@ class test_stdlib(TestCore):
 
         self.assert_has_error(self.run_script('''
         print 'hello ' + $name;
-        '''))
+        '''), 'RuntimeError')
 
         program = self.run_script('''
         raise ['SomeError', 'this is error'];
         ''')
-        self.assert_has_error(program)
-        self.assert_equals(program['runtime_error'][0], 'SomeError')
-        self.assert_equals(program['runtime_error'][1], 'this is error')
+        self.assert_has_error(program, 'SomeError')
+        self.assert_equals(program['runtime_error']['type'], 'SomeError')
+        self.assert_equals(program['runtime_error']['message'], 'this is error')
 
         self.assert_has_error(self.run_script('''
         assert 2 == 3;
-        '''))
+        '''), 'AssertError')
 
         self.assert_has_error(self.run_script('''
         assert 1 > 10;
-        '''))
+        '''), 'AssertError')
 
         self.assert_has_error(self.run_script('''
         assert False;
-        '''))
+        '''), 'AssertError')
 
         self.assert_has_error(self.run_script('''
         set $age; mem 18; copy $age;
         assert $age > 30;
-        '''))
+        '''), 'AssertError')
 
         self.run_without_error('''
         assert True;