+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Pashmak Changelog
 
+## v1.3.1 (2020-10-09)
+
+#### Fixed
+- fixed include system file path bug [(f056d194)](https://github.com/parsampsh/pashmak/commit/f056d19451adf32e13ab80901de7114166499cc8)
+- clean up and optimize code
+
 ## v1.3 (2020-10-09)
 
 #### Fixed